requirements.txt
openai
gtts
pyaudio
SpeechRecognition
python-dotenv
playsound